Split and typo
